+++ src/constant/paths.ts
@@ -2,16 +2,8 @@ export const INDEX_PATH_NAME = "";
 export const HOME_PATH_NAME = "home";
 export const REGISTRATION_PATH_NAME = "registration";
 export const SETTINGS_PATH_NAME = "settings";
-export const PROFILE_PATH_NAME = "profile";
-export const TEAMS_PATH_NAME = "teams";
-export const SHARE_PATH_NAME = "share";
-export const TAGS_PATH_NAME = "tags";
 
 export const INDEX_PATH = `/${INDEX_PATH_NAME}`;
 export const HOME_PATH = `/${HOME_PATH_NAME}`;
 export const REGISTRATION_PATH = `/${REGISTRATION_PATH_NAME}`;
 export const SETTINGS_PATH = `/${SETTINGS_PATH_NAME}`;
-export const PROFILE_PATH = `/${SETTINGS_PATH_NAME}/${PROFILE_PATH_NAME}`;
-export const TEAMS_PATH = `/${SETTINGS_PATH_NAME}/${TEAMS_PATH_NAME}`;
-export const SHARE_PATH = `/${SETTINGS_PATH_NAME}/${SHARE_PATH_NAME}`;
-export const TAGS_PATH = `/${SETTINGS_PATH_NAME}/${TAGS_PATH_NAME}`;
